requirement_dis.txt
annotated-types==0.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
diffusers==0.17.0
eval_type_backport==0.2.2
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.13.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lmdb==1.7.2
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.3.0
pip==25.1
platformdirs==4.3.8
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.13.1
sentry-sdk==2.33.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
wandb==0.21.0
wheel==0.45.1
zipp==3.23.0
